+++ client/src/components/navBar.js
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import PropTypes from 'prop-types';
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Divider from '@mui/material/Divider';
@@ -14,13 +13,25 @@ import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 import Button from '@mui/material/Button';
 
-import { useQuery } from '@apollo/client'
-import { GET_FREELANCERS } from '../utils/queries'
+import { useState, useEffect } from 'react';
+
+// Session Authenticator
+import Auth from '../utils/Auth';
 
 const drawerWidth = 240;
-const navItems = ['Home', 'Sign In', 'Sign Up'];
+// let navItems = ['Home', 'Sign In', 'Sign Up'];
 
 function NavBar (props) {
+
+    const [navItems, setNavItems] = useState([]);
+
+    useEffect(() => {
+        const isLogged = Auth.loggedIn();
+        console.log(isLogged);
+        if (!isLogged) setNavItems(['Home', 'Sign In', 'Sign Up'])
+        if (isLogged) setNavItems(['Home', 'Sign Out']);
+    }, []);
+
 
     const { window } = props;
     const [mobileOpen, setMobileOpen] = React.useState(false);
