+++ client/src/pages/LandingPage.js
@@ -1,15 +1,12 @@
 import React from 'react';
-import { Grid, Button, Container } from '@mui/material';
+import { Button, Container } from '@mui/material';
 import Typography from '@mui/material/Typography';
-import Paper from '@mui/material/Paper';
-import TextField from '@mui/material/TextField';
+import NavBar from '../components/navBar';
 
 
 export default function LandingPage() {
     return (
-    <Container fixed
-    text align= "left"
-    >
+      <Container fixed text align= "left">
         <Typography variant="h2" gutterBottom>Start to Dev-connect!</Typography>
           <Typography variant="h5" gutterBottom>
             Welcome to the best professional network where you can find the best developers
